# contracts/commerce/openapi-internal.yaml
openapi: 3.1.0
info:
  title: Commerce — API interna
  version: 1.3.0
  description: 'Contrato **vigente** consolidado. Fronteira: `bff-admin` → `commerce` (área financeira, Catálogo,
    cortesias), `bff-student` → `commerce` (vitrine) e serviços → `commerce` (decisão de acesso).


    Fonte da verdade do contrato desta fronteira. Mudanças entram por PR neste arquivo, promovidas a partir do recorte
    do PRD (ver `contracts/README.md`). Os documentos de origem, arquivados, são só histórico:


    - `tasks/archive/prd-acesso-interno/internal-api-contract-commerce.yaml` 1.0.0

    - `tasks/archive/prd-vitrine-oferta/internal-api-contract-commerce.yaml` 1.1.0

    - `tasks/archive/prd-concessao-acesso/internal-api-contract-commerce.yaml` 1.2.0

    - `tasks/archive/prd-progresso-aluno/internal-api-contract-commerce.yaml` 1.3.0


    **Notas de prd-acesso-interno (1.0.0):** Recorte interno de CAP-002 (TechSpec aprovada em 2026-09-25, V-07;
    ADR-0005). Consumidor único conhecido: `bff-admin`. `commerce` valida localmente o JWT de usuário emitido por
    Identity (assinatura pelo JWKS de Identity, emissor `identity`, audiência `commerce`, validade) e decide pela
    permissão do claim `permissions`, sem consultar Identity e sem confiar no BFF. A área não tem dado até CAP-011.


    **Notas de prd-vitrine-oferta (1.1.0):** Recorte interno do primeiro PRD de CAP-003 (`tasks/archive/prd-vitrine-oferta/prd.md`
...
    vem do `tenantId` do token ou da asserção (G07).


...


...

    '
...
servers:
...
paths:
...
  /orders:
    post:
      tags:
      - Compra
      summary: Criar o pedido de uma oferta
      description: 'Cria o pedido em `awaiting-payment` com curso, opção, preço e vigência da oferta **congelados
        neste instante** (RF-03, RN-V03, RN-V04). Só conta de aluno (RN-V01): token de ator interno não tem o escopo
        de aluno e é recusado com 403.


        - Já existe pedido do aluno aguardando pagamento **para a mesma oferta** → 200 com **esse** pedido, inalterado,
        sem criar outro (RN-V06). Pedido pendente de outra oferta do mesmo curso não impede (RF-03, quinto critério).

        - Oferta não publicada, inexistente ou de outra escola → 404, nada criado (RN-V02).

        - `Idempotency-Key` obrigatória (G09): repetição com o mesmo corpo devolve o mesmo resultado e o mesmo status;
        corpo diferente → 422 `IDEMPOTENCY_KEY_REUSED`.


        Publica `vendas.pedido-criado.v1` pelo outbox, na mesma transação (G06). Não abre pagamento: isso é `startOrderPaymentInternal`.

        '
      operationId: createOrderInternal
      security:
      - StudentUserToken: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderInput'
            examples:
              dozeMeses:
                value:
                  offerId: 7c8d9e0f-1a2b-4c3d-8e4f-5a6b7c8d9e0f
      responses:
        '201':
          description: Pedido criado.
          headers:
            Location:
              description: Endereço do pedido criado.
              schema:
                type: string
                examples:
                - /internal/v1/orders/1b2c3d4e-5f60-4718-9a2b-3c4d5e6f7a81
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentOrder'
              examples:
                criado:
                  $ref: '#/components/examples/OrderAwaitingNoMethod'
        '200':
          description: Já havia pedido do aluno aguardando pagamento nesta oferta; é ele que volta (RN-V06).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentOrder'
              examples:
                pendente:
                  $ref: '#/components/examples/OrderAwaitingBoleto'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/TokenInvalid'
        '403':
          $ref: '#/components/responses/StudentScopeDenied'
        '404':
          $ref: '#/components/responses/OfferNotAvailable'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
    get:
      tags:
      - Meus pedidos
      summary: Listar os pedidos do aluno
      description: Pedidos do aluno do token, na escola do token, do mais recente para o mais antigo (RF-10). Mostram
        o que o pedido congelou, mesmo que a oferta tenha mudado ou saído da vitrine (RN-V03, RN-V12). Sem pedidos
        → lista vazia.
      operationId: listStudentOrdersInternal
      security:
      - StudentUserToken: []
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Página de pedidos do aluno.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentOrderPage'
              examples:
                tres:
                  value:
                    data:
                    - orderId: 2c3d4e5f-6071-4829-8b3c-4d5e6f7a8b92
                      number: '000124'
                      status: awaiting-payment
                      course:
                        courseId: 9a7e5c31-2b4d-4f68-8e01-6d3c5a7b9f24
                        title: React na prática
                      offer:
                        offerId: 5a6b7c8d-9e0f-4a1b-8c2d-3e4f5a6b7c8d
                        name: Acesso por 6 meses
                      priceCents: 29700
                      currency: BRL
                      accessPeriod:
                        type: months
                        months: 6
                      paymentMethod: boleto
                      pendingPayment:
                        method: boleto
                        expiresAt: '2026-10-09T02:59:59Z'
                      paymentPageExpiresAt: null
                      accessGrantedAt: null
                      createdAt: '2026-10-06T10:00:00Z'
                      paidAt: null
                      expiredAt: null
                      cancelledAt: null
                    - orderId: 1b2c3d4e-5f60-4718-9a2b-3c4d5e6f7a81
                      number: '000123'
                      status: paid
                      course:
                        courseId: 6f1e2d3c-4b5a-4c69-8d7e-1f2a3b4c5d6e
                        title: .NET do zero à API
                      offer:
                        offerId: 7c8d9e0f-1a2b-4c3d-8e4f-5a6b7c8d9e0f
                        name: Acesso por 12 meses
                      priceCents: 49700
                      currency: BRL
                      accessPeriod:
                        type: months
                        months: 12
                      paymentMethod: card
                      pendingPayment: null
                      paymentPageExpiresAt: null
                      accessGrantedAt: '2026-10-05T14:21:04Z'
                      createdAt: '2026-10-05T14:19:30Z'
                      paidAt: '2026-10-05T14:21:02Z'
                      expiredAt: null
                      cancelledAt: null
                    pagination:
                      page: 1
                      size: 10
                      total: 2
                      totalPages: 1
                vazio:
                  value:
                    data: []
                    pagination:
                      page: 1
                      size: 10
                      total: 0
                      totalPages: 0
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/TokenInvalid'
        '403':
          $ref: '#/components/responses/StudentScopeDenied'
  /orders/{orderId}:
    parameters:
    - $ref: '#/components/parameters/OrderId'
    get:
      tags:
      - Meus pedidos
      summary: Obter um pedido do aluno
      description: 'A situação **real** do pedido, para a página de retorno e para a retomada (RF-05, RF-06, RN-CB04).
        O cliente consulta de novo até a situação mudar; a resposta não é guardada (`no-store`).


        - `paid` → pago; `accessGrantedAt` preenchido quando a concessão existe (RF-05: *Ir para o curso*).

        - `awaiting-payment` com `pendingPayment` → PIX ou boleto gerado, ainda não pago.

        - `awaiting-payment` sem `pendingPayment` → página de pagamento aberta (`paymentPageExpiresAt`) ou ainda
        não aberta (nulo). O cliente sabe se o aluno concluiu ou saiu pela própria volta da página; "Confirmando"
        é `awaiting-payment` sem `pendingPayment` depois de o aluno ter concluído.

        - `expired`, `cancelled` → terminais sem concessão, salvo confirmação tardia (RN-V08), que muda para `paid`.


        Pedido de outro aluno, de outra escola ou inexistente → 404 indistinto.

        '
      operationId: getStudentOrderInternal
      security:
      - StudentUserToken: []
      responses:
        '200':
          description: O pedido.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentOrder'
              examples:
                pago:
                  $ref: '#/components/examples/OrderPaid'
                boleto:
                  $ref: '#/components/examples/OrderAwaitingBoleto'
                semMeio:
                  $ref: '#/components/examples/OrderAwaitingNoMethod'
        '401':
          $ref: '#/components/responses/TokenInvalid'
        '403':
          $ref: '#/components/responses/StudentScopeDenied'
        '404':
          $ref: '#/components/responses/OrderNotFound'
...
  /finance/orders:
    get:
      tags:
      - Pedidos no backoffice
      summary: Listar os pedidos da escola
      description: 'Pedidos da escola do token, mais recentes primeiro (RF-11). Exige `financeiro.ler` (DP-09); só
        leitura. Filtros combináveis: `status`, `courseId`, `studentId` e período de **criação** do pedido (`createdFrom`
        e `createdTo`, datas inclusivas no fuso da escola). O aluno vai por referência (`studentId`): nome e e-mail
        são resolvidos pelo `bff-admin` em Identity, nunca guardados aqui.
  
        '
      operationId: listFinanceOrdersInternal
      security:
      - StaffUserToken: []
      parameters:
      - name: status
        in: query
        required: false
        description: Situação do pedido. Ausente = todas.
        schema:
          $ref: '#/components/schemas/OrderStatus'
      - name: courseId
        in: query
        required: false
        description: Curso do pedido.
        schema:
          type: string
          format: uuid
      - name: studentId
        in: query
        required: false
        description: Conta de aluno em Identidade e Acesso.
        schema:
          type: string
          format: uuid
      - name: createdFrom
        in: query
        required: false
        description: Primeiro dia do período, inclusivo, no fuso da escola.
        schema:
          type: string
          format: date
        examples:
          inicio:
            value: '2026-10-01'
      - name: createdTo
        in: query
        required: false
        description: Último dia do período, inclusivo, no fuso da escola. Anterior a `createdFrom` → 400.
        schema:
          type: string
          format: date
        examples:
          fim:
            value: '2026-10-31'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Página de pedidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceOrderPage'
              examples:
                pendentes:
                  value:
                    data:
                    - orderId: 2c3d4e5f-6071-4829-8b3c-4d5e6f7a8b92
                      number: '000124'
                      studentId: 2b3c4d5e-6f70-4a81-9b92-a3b4c5d6e7f8
                      status: awaiting-payment
                      courseId: 9a7e5c31-2b4d-4f68-8e01-6d3c5a7b9f24
                      courseTitle: React na prática
                      offerName: Acesso por 6 meses
                      priceCents: 29700
                      currency: BRL
                      paymentMethod: boleto
                      createdAt: '2026-10-06T10:00:00Z'
                      paidAt: null
                    pagination:
                      page: 1
                      size: 10
                      total: 1
                      totalPages: 1
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/TokenInvalid'
        '403':
          $ref: '#/components/responses/FinancePermissionDenied'
  /finance/orders/{orderId}:
    parameters:
    - $ref: '#/components/parameters/OrderId'
    get:
      tags:
      - Pedidos no backoffice
      summary: Obter o detalhe de um pedido
      description: Tudo o que a lista mostra, mais a vigência congelada, os momentos de cada mudança, a referência
        do pagamento no gateway (para localizar no painel dele) e a concessão correspondente, quando existe (RF-11).
        Exige `financeiro.ler`. Pedido de outra escola ou inexistente → 404 indistinto.
      operationId: getFinanceOrderInternal
      security:
      - StaffUserToken: []
      responses:
        '200':
          description: O pedido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceOrderDetail'
              examples:
                pago:
                  value:
                    orderId: 1b2c3d4e-5f60-4718-9a2b-3c4d5e6f7a81
                    number: '000123'
                    studentId: 9d8c7b6a-5f4e-4d3c-8b2a-1f0e9d8c7b6a
                    status: paid
                    courseId: 6f1e2d3c-4b5a-4c69-8d7e-1f2a3b4c5d6e
                    courseTitle: .NET do zero à API
                    offerId: 7c8d9e0f-1a2b-4c3d-8e4f-5a6b7c8d9e0f
                    offerName: Acesso por 12 meses
                    priceCents: 49700
                    currency: BRL
                    accessPeriod:
                      type: months
                      months: 12
                    paymentMethod: card
                    paymentReference: pi_3Q2w3E4r5T6y7U8i0
                    paidAmountCents: 49700
                    grantId: 4c5d6e7f-8091-4a2b-9c3d-4e5f6a7b8c9d
                    accessGrantedAt: '2026-10-05T14:21:04Z'
                    createdAt: '2026-10-05T14:19:30Z'
                    paymentPageExpiresAt: '2026-10-06T14:19:41Z'
                    paidAt: '2026-10-05T14:21:02Z'
                    expiredAt: null
                    cancelledAt: null
        '401':
          $ref: '#/components/responses/TokenInvalid'
        '403':
          $ref: '#/components/responses/FinancePermissionDenied'
        '404':
          $ref: '#/components/responses/OrderNotFound'
components:
  securitySchemes:
    StudentUserToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT RS256 de aluno emitido por Identity (`validateStudentSessionInternal` com `audience: commerce`),
        vida de 5 minutos, validado localmente pelo JWKS de Identity (ADR-0013, estendida a `commerce` por C-03;
        ADR própria na TechSpec). Política de aluno: `scope` igual a `orders:use` **e ausência da claim `permissions`**.
        Claims `iss`, `aud`, `sub` (= `studentId`), `tenantId`, `sessionId`, `scope`, `exp`, `jti`. **Sem `email`**:
        a claim só existe na audiência `media`.'
    StaffUserToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT RS256 de ator interno emitido por Identity (`validateStaffSessionInternal` com `audience:
        commerce`), vida curta, claims iss, aud, sub, tenantId, sessionId, scope, roles, permissions, exp, jti.
        Chave pública por `kid` no JWKS de Identity. As operações de Cortesias exigem `cortesia.conceder`.'
    StudentBffServiceAssertion:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Asserção de serviço assimétrica e curta assinada pela chave exclusiva do `bff-student` (mecanismo
        da ADR-0004, estendido a `commerce` por C-04; ADR própria na TechSpec). `commerce` valida assinatura, `kid`,
        emissor `bff-student`, audiência `commerce`, escopo, `tenantId`, validade e `jti` contra replay. Escopos:
        `showcase:read` nas leituras e `purchase-intent:write` no clique. Não representa aluno nem visitante.'
    DomainServiceAssertion:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Asserção de serviço assimétrica e curta assinada pela chave exclusiva do serviço chamador (`media`
        ou `learning`), no mecanismo da ADR-0004 estendido por ADR-0009 (emissores: ADR-0012, revisada pela ADR-0015).
        `commerce` valida assinatura, `kid`, emissor, audiência `commerce`, o escopo da rota (ADR-0015: `access-decision:read`
        em `decideAccessInternal`, para `media` e `learning`; `course-access:read` em `listStudentCourseAccessInternal`,
        só para `learning`), `tenantId` permitido ao emissor, validade e `jti` contra replay. Não representa aluno.'
  schemas:
    StudentOrderPage:
      type: object
      additionalProperties: false
      required:
      - data
      - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/StudentOrder'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
    PriceCents:
      type: integer
      minimum: 1
      maximum: 9999999
      description: Preço em centavos de real (BRL), de R$ 0,01 a R$ 99.999,99 (DP-03). Sem parcelamento.
      examples:
      - 49700
    AccessPeriod:
      description: 'Vigência da concessão, na forma única de `AccessPeriod` de CAP-003 (C-05): por período, em meses
        inteiros de 1 a 60 (DP-04), contados a partir do momento da concessão (RN-D04), ou vitalícia.

        '
      oneOf:
      - type: object
        additionalProperties: false
        required:
        - type
        - months
        properties:
          type:
            const: months
          months:
            type: integer
            minimum: 1
            maximum: 60
            examples:
            - 12
      - type: object
        additionalProperties: false
        required:
        - type
        properties:
          type:
            const: lifetime
...
    StudentOrder:
      type: object
      additionalProperties: false
      required:
      - orderId
      - number
      - status
      - course
      - offer
      - priceCents
      - currency
      - accessPeriod
      - paymentMethod
      - pendingPayment
      - paymentPageExpiresAt
      - accessGrantedAt
      - createdAt
      - paidAt
      - expiredAt
      - cancelledAt
      properties:
        orderId:
          type: string
          format: uuid
        number:
          type: string
          pattern: ^[0-9]{6,}$
          description: Número do pedido para o aluno e o financeiro ("Pedido nº 000123"), sequencial por escola.
            Também no comprovante.
          examples:
          - '000123'
        status:
          $ref: '#/components/schemas/OrderStatus'
        course:
          $ref: '#/components/schemas/CourseRef'
        offer:
          $ref: '#/components/schemas/OfferRef'
        priceCents:
          $ref: '#/components/schemas/PriceCents'
        currency:
          $ref: '#/components/schemas/Currency'
        accessPeriod:
          $ref: '#/components/schemas/AccessPeriod'
        paymentMethod:
          oneOf:
          - $ref: '#/components/schemas/PaymentMethod'
          - type: 'null'
          description: Meio do PIX ou boleto pendente, ou o que pagou. Nulo enquanto nenhum foi escolhido.
        pendingPayment:
          oneOf:
          - $ref: '#/components/schemas/PendingPayment'
          - type: 'null'
        paymentPageExpiresAt:
          type:
          - string
          - 'null'
          format: date-time
          description: Até quando vale a página de pagamento aberta sem meio escolhido (24 h, RN-CB03). Nulo se
            nunca foi aberta, depois de escolhido o meio, ou em pedido terminal.
        accessGrantedAt:
          type:
          - string
          - 'null'
          format: date-time
          description: Momento da concessão deste pedido em Matrícula, quando já existe (C-07). **Exibição**, não
            decisão de acesso (RN-V10).
        createdAt:
          type: string
          format: date-time
        paidAt:
          type:
          - string
          - 'null'
          format: date-time
          description: Momento em que o gateway confirmou o pagamento.
        expiredAt:
          type:
          - string
          - 'null'
          format: date-time
        cancelledAt:
          type:
          - string
          - 'null'
          format: date-time
...
    PaymentMethod:
      type: string
      enum:
      - card
      - pix
      - boleto
      description: Meio de pagamento (RN-CB02, RN-CB05). Clientes devem tolerar valor novo.
    Currency:
      type: string
      const: BRL
    OfferRef:
      type: object
      additionalProperties: false
      required:
      - offerId
      - name
      properties:
        offerId:
          type: string
          format: uuid
        name:
          type: string
          description: Nome da opção de compra. No pedido, o congelado na criação.
    CourseRef:
      type: object
      additionalProperties: false
      required:
      - courseId
      - title
      properties:
        courseId:
          type: string
          format: uuid
        title:
          type: string
          description: Título do curso. No pedido, o congelado na criação (RN-V03).
    OrderStatus:
      type: string
      enum:
      - awaiting-payment
      - paid
      - expired
      - cancelled
      description: Situação do pedido (RN-V05). Só `paid` gera concessão. `paid` é terminal; `expired` e `cancelled`
        só mudam para `paid` por confirmação tardia (RN-V08). Clientes devem tolerar valor novo.
    CreateOrderInput:
      type: object
      additionalProperties: false
      required:
      - offerId
      properties:
        offerId:
          type: string
          format: uuid
          description: Oferta publicada a comprar. Preço e vigência **não** são aceitos do cliente; vêm da oferta
            no instante da criação.
...
    FinanceOrderPage:
      type: object
      additionalProperties: false
      required:
      - data
      - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FinanceOrderSummary'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 'Chave opaca única por intenção de escrita (G09). Repetição com o mesmo corpo devolve o mesmo
        resultado na janela de 24 horas; corpo diferente com a mesma chave → 422 `IDEMPOTENCY_KEY_REUSED`.

        '
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    Page:
      name: _page
      in: query
      required: false
      description: Página, a partir de 1.
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: _size
      in: query
      required: false
      description: Itens por página.
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 10
...
    OrderId:
      name: orderId
      in: path
      required: true
      description: Identificador do pedido.
      schema:
        type: string
        format: uuid
  responses:
    InvalidRequest:
      description: Requisição malformada.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            malformada:
              value:
                type: about:blank
                title: Requisição inválida.
                status: 400
                code: INVALID_REQUEST
                traceId: 00-9f2b5d6c4e7a8b90123456789abcdef0-0123456789abcdef-01
...
    OfferNotAvailable:
      description: Oferta não publicada, inexistente ou de outra escola — indistintas. O clique não é contado.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            indisponivel:
              value:
                type: about:blank
                title: Oferta não disponível.
                status: 404
                code: OFFER_NOT_AVAILABLE
                traceId: 00-9f2b5d6c4e7a8b90123456789abcdef0-0123456789abcdef-01
    TokenInvalid:
      description: JWT ausente, expirado, de audiência errada ou com assinatura inválida; code `TOKEN_INVALID`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            e:
              value:
                type: about:blank
                title: Token de acesso inválido.
                status: 401
                code: TOKEN_INVALID
                traceId: 00-9f2b5d6c4e7a8b90123456789abcdef0-0123456789abcdef-01
            token:
              value:
                type: about:blank
                title: Token inválido.
                status: 401
                code: TOKEN_INVALID
                traceId: 00-9f2b5d6c4e7a8b90123456789abcdef0-0123456789abcdef-01
...
    OrderNotFound:
      description: Pedido inexistente, de outro aluno ou de outra escola — indistintos; code `ORDER_NOT_FOUND`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            naoEncontrado:
              value:
                type: about:blank
                title: Pedido não encontrado.
                status: 404
                code: ORDER_NOT_FOUND
                traceId: 00-9f2b5d6c4e7a8b90123456789abcdef0-0123456789abcdef-01
    StudentScopeDenied:
      description: Token válido sem a política de aluno — sem `orders:use` ou com `permissions` (token de ator interno,
        RN-V01); code `SCOPE_DENIED`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            atorInterno:
              value:
                type: about:blank
                title: Escopo insuficiente.
                status: 403
                code: SCOPE_DENIED
                traceId: 00-9f2b5d6c4e7a8b90123456789abcdef0-0123456789abcdef-01
    IdempotencyKeyReused:
      description: Mesma `Idempotency-Key` com corpo diferente; code `IDEMPOTENCY_KEY_REUSED`. Nada é criado.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            reuso:
              value:
                type: about:blank
                title: Chave de idempotência já usada com outro conteúdo.
                status: 422
                code: IDEMPOTENCY_KEY_REUSED
                traceId: 00-9f2b5d6c4e7a8b90123456789abcdef0-0123456789abcdef-01
...
    FinancePermissionDenied:
      description: Token de ator sem `financeiro.ler`, ou token de aluno; code `PERMISSION_DENIED`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            semPermissao:
              value:
                type: about:blank
                title: Sem permissão para esta área.
                status: 403
                code: PERMISSION_DENIED
                traceId: 00-9f2b5d6c4e7a8b90123456789abcdef0-0123456789abcdef-01
  examples:
...
    OrderAwaitingNoMethod:
      summary: Pedido criado, página de pagamento ainda não aberta
      value:
        orderId: 1b2c3d4e-5f60-4718-9a2b-3c4d5e6f7a81
        number: '000123'
        status: awaiting-payment
        course:
          courseId: 6f1e2d3c-4b5a-4c69-8d7e-1f2a3b4c5d6e
          title: .NET do zero à API
        offer:
          offerId: 7c8d9e0f-1a2b-4c3d-8e4f-5a6b7c8d9e0f
          name: Acesso por 12 meses
        priceCents: 49700
        currency: BRL
        accessPeriod:
          type: months
          months: 12
        paymentMethod: null
        pendingPayment: null
        paymentPageExpiresAt: null
        accessGrantedAt: null
        createdAt: '2026-10-05T14:19:30Z'
        paidAt: null
        expiredAt: null
        cancelledAt: null
    OrderAwaitingBoleto:
      summary: Boleto gerado, ainda não pago
      value:
        orderId: 2c3d4e5f-6071-4829-8b3c-4d5e6f7a8b92
        number: '000124'
        status: awaiting-payment
        course:
          courseId: 9a7e5c31-2b4d-4f68-8e01-6d3c5a7b9f24
          title: React na prática
        offer:
          offerId: 5a6b7c8d-9e0f-4a1b-8c2d-3e4f5a6b7c8d
          name: Acesso por 6 meses
        priceCents: 29700
        currency: BRL
        accessPeriod:
          type: months
          months: 6
        paymentMethod: boleto
        pendingPayment:
          method: boleto
          expiresAt: '2026-10-09T02:59:59Z'
        paymentPageExpiresAt: null
        accessGrantedAt: null
        createdAt: '2026-10-06T10:00:00Z'
        paidAt: null
        expiredAt: null
        cancelledAt: null
    OrderPaid:
      summary: Pago por cartão, concessão já existente
      value:
        orderId: 1b2c3d4e-5f60-4718-9a2b-3c4d5e6f7a81
        number: '000123'
        status: paid
        course:
          courseId: 6f1e2d3c-4b5a-4c69-8d7e-1f2a3b4c5d6e
          title: .NET do zero à API
        offer:
          offerId: 7c8d9e0f-1a2b-4c3d-8e4f-5a6b7c8d9e0f
          name: Acesso por 12 meses
        priceCents: 49700
        currency: BRL
        accessPeriod:
          type: months
          months: 12
        paymentMethod: card
        pendingPayment: null
        paymentPageExpiresAt: null
        accessGrantedAt: '2026-10-05T14:21:04Z'
        createdAt: '2026-10-05T14:19:30Z'
        paidAt: '2026-10-05T14:21:02Z'
        expiredAt: null
        cancelledAt: null
...
  headers:
    NoStore:
      description: '`private, no-store`: a situação do pedido muda com o tempo e não é guardada pelo consumidor.'
      schema:
        type: string
        examples:
        - private, no-store